feat: 코드 에디터 컴포넌트 분리
